feat: fix vimeo video integration after upload/create asset
