+++ .env.example
@@ -46,6 +46,7 @@ PG_PASSWORD=${DB_PASSWORD}
 PG_DB_NAME=${DB_NAME}
 CACHE_VIEWS=
 
+VIMEO_URL=https://vimeo.com
 VIMEO_CLIENT_ID=f6ac6453c4f70ff4713dcbd6f67cda4692f78b66
 VIMEO_CLIENT_SECRET=XZfQjM5T1UfHZiymiBPXDO4XIGS4RXBeD9g6UIeapIQ9gC5Y+TH+NaUew/AsPokq8ACZZU7v6yl4VeGu3Ebt7oWi6NS0jIQVhS4rR8KDf0H6Rodh9KJLGg4YcSJUyyV6
-VIMEO_ACCESS_TOKEN=95c265579f74560199e7a73549212d2e
+VIMEO_ACCESS_TOKEN=6dfb64781647e875dcf0095ad67387be
