+++ app/Models/Asset.ts
@@ -16,4 +16,5 @@ export default class Asset extends BaseModel {
   @column() public asset: string
   @column() public owner: string
   @column() public owner_id: number
+  @column() public vimeo_url: string
 }
